librustdoc: Unconfigure tests during normal build
